@@ -5,12 +5,12 @@ repository: https://github.com/mateusfccp/update_available/tree/master/update_av
 publish_to: none
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
-  flutter: ">=1.20.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   update_available:
     path: ..
 
 dev_dependencies:
-  pedantic: ^1.9.0
+  flutter_lints: ^1.0.4
