@@ -1,4 +1,4 @@
 import 'dart:ui';
 
-const Green = Color(0xFFA1DF31);
-const Red = Color(0xFFDE3B1B);
+const green = Color(0xFFA1DF31);
+const red = Color(0xFFDE3B1B);
