@@ -6,18 +6,20 @@ import 'json.dart';
 
 const _lookupURL = 'https://itunes.apple.com/lookup';
 
-final GetIOSPackageVersion httpGetIOSPackageVersion = (String bundleId) async {
-  final uri = Uri.parse('$_lookupURL?bundleId=$bundleId');
-  final response = await get(uri);
-  final versionString = getStringByKey(response.body)('version');
+GetIOSPackageVersion httpGetIOSPackageVersion() {
+  return (String bundleId) async {
+    final uri = Uri.parse('$_lookupURL?bundleId=$bundleId');
+    final response = await get(uri);
+    final versionString = getStringByKey(response.body)('version');
 
-  if (versionString == null) {
-    return null;
-  }
+    if (versionString == null) {
+      return null;
+    }
 
-  try {
-    return Version.parse(versionString);
-  } on FormatException {
-    return null;
-  }
-};
+    try {
+      return Version.parse(versionString);
+    } on FormatException {
+      return null;
+    }
+  };
+}
